# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 7of1 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /oauth/{platform}/connect:
    get:
      operationId: oauthConnect
      tags: [oauth]
      summary: Initiate OAuth connection
      description: >
        Redirects the creator to the external OAuth authorization page for the
        specified platform (17live, line, youtube). Requires creator auth session.
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OAuthPlatform"
      responses:
        "302":
          description: Redirect to external OAuth authorization page
        "400":
          description: Unsupported platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /oauth/{platform}/callback:
    get:
      operationId: oauthCallback
      tags: [oauth]
      summary: OAuth callback handler
      description: >
        Handles the redirect from the external OAuth provider. Verifies state,
        exchanges code for tokens, stores encrypted tokens, redirects to dashboard.
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OAuthPlatform"
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to dashboard (with oauth_connected or oauth_error query param)
        "400":
          description: Invalid state or missing parameters
        "503":
          description: Database not configured
  /oauth/{platform}/status:
    get:
      operationId: oauthStatus
      tags: [oauth]
      summary: Get OAuth connection status
      description: Returns connection status for the platform (no token values exposed). Requires creator auth.
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OAuthPlatform"
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthStatusResult"
        "400":
          description: Unsupported platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /oauth/{platform}:
    delete:
      operationId: oauthRevoke
      tags: [oauth]
      summary: Revoke OAuth connection
      description: Revokes and deletes stored tokens for the platform. Requires creator auth.
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OAuthPlatform"
      responses:
        "200":
          description: Connection revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthRevokeResult"
        "400":
          description: Unsupported platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Webhooks ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        remainingBalance:
          type: ["integer", "null"]
      required:
        - error
...
    OAuthPlatform:
      type: string
      enum: ["17live", "line", "youtube"]
    OAuthStatusResult:
      type: object
      properties:
        platform:
          $ref: "#/components/schemas/OAuthPlatform"
        connected:
          type: boolean
        expiresAt:
          type: ["string", "null"]
          format: date-time
        scope:
          type: ["string", "null"]
      required:
        - platform
        - connected
    OAuthRevokeResult:
      type: object
      properties:
        revoked:
          type: boolean
        platform:
          $ref: "#/components/schemas/OAuthPlatform"
      required:
        - revoked
        - platform

    # ── Refunds ──────────────────────────────────────────────────────────────
...