# lib/api-zod/src/generated/types/oAuthPlatform.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 7of1 API specification
 * OpenAPI spec version: 0.1.0
 */

export type OAuthPlatform = typeof OAuthPlatform[keyof typeof OAuthPlatform];


export const OAuthPlatform = {
  '17live': '17live',
  line: 'line',
  youtube: 'youtube',
} as const;